[feat]: #79 기존 XML 코드 -> Compose 로 마이그레이션
