+++ app/src/main/res/navigation/nav_graph.xml
@@ -53,23 +53,6 @@
     </fragment>
 
     <fragment
-        android:id="@+id/fragment_report"
-        android:name="com.example.findu.presentation.ui.report.ReportFragment"
-        android:label="Report"
-        tools:layout="@layout/fragment_report">
-        <action
-            android:id="@+id/action_fragment_report_to_fragment_witness_report"
-            app:destination="@id/fragment_witness_report"
-            app:popUpTo="@id/fragment_report"
-            app:popUpToInclusive="true" />
-        <action
-            android:id="@+id/action_fragment_report_to_fragment_missing_report"
-            app:destination="@id/fragment_missing_report"
-            app:popUpTo="@id/fragment_report"
-            app:popUpToInclusive="true" />
-    </fragment>
-
-    <fragment
         android:id="@+id/fragment_info"
         android:name="com.example.findu.presentation.ui.info.InfoFragment"
         android:label="Info"
@@ -112,44 +95,6 @@
         <action
             android:id="@+id/action_fragment_my_to_fragment_my_keep_animals"
             app:destination="@id/fragment_my_keep_animals" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/fragment_witness_report"
-        android:name="com.example.findu.presentation.ui.report.WitnessReportFragment"
-        android:label="Witness Report"
-        tools:layout="@layout/fragment_witness_report">
-        <action
-            android:id="@+id/action_fragment_witness_report_to_fragment_home"
-            app:destination="@id/fragment_home"
-            app:popUpToInclusive="true" />
-        <action
-            android:id="@+id/action_fragment_witness_report_to_fragment_report_camera"
-            app:destination="@id/fragment_report_camera" />
-        <action
-            android:id="@+id/action_fragment_witness_report_to_fragment_my_report_history"
-            app:destination="@id/fragment_my_report_history" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/fragment_missing_report"
-        android:name="com.example.findu.presentation.ui.report.MissingReportFragment"
-        android:label="Missing Report"
-        tools:layout="@layout/fragment_missing_report">
-        <action
-            android:id="@+id/action_fragment_missing_report_to_fragment_home"
-            app:destination="@id/fragment_home"
-            app:popUpToInclusive="true" />
-        <action
-            android:id="@+id/action_fragment_missing_report_to_fragment_new_missing_report"
-            app:destination="@id/fragment_new_missing_report"
-            app:popUpToInclusive="true" />
-        <action
-            android:id="@+id/action_fragment_missing_report_to_fragment_report_camera"
-            app:destination="@id/fragment_report_camera" />
-        <action
-            android:id="@+id/action_fragment_missing_report_to_fragment_my_report_history"
-            app:destination="@id/fragment_my_report_history" />
     </fragment>
 
     <fragment
@@ -264,36 +209,36 @@
     </fragment>
 
     <fragment
-        android:id="@+id/fragment_new_missing_report"
-        android:name="com.example.findu.presentation.ui.report.missing.NewMissingReportFragment"
-        android:label="New Missing Report"
-        tools:layout="@layout/fragment_new_missing_report">
+        android:id="@+id/fragment_missing_report"
+        android:name="com.example.findu.presentation.ui.report.missing.MissingReportFragment"
+        android:label="Missing Report"
+        tools:layout="@layout/fragment_missing_report">
         <action
-            android:id="@+id/action_fragment_new_missing_report_to_fragment_home"
+            android:id="@+id/action_fragment_missing_report_to_fragment_home"
             app:destination="@id/fragment_home"
             app:popUpToInclusive="true" />
         <action
-            android:id="@+id/action_fragment_new_missing_report_to_fragment_report_camera"
+            android:id="@+id/action_fragment_missing_report_to_fragment_report_camera"
             app:destination="@id/fragment_report_camera" />
         <action
-            android:id="@+id/action_fragment_new_missing_report_to_fragment_my_report_history"
+            android:id="@+id/action_fragment_missing_report_to_fragment_my_report_history"
             app:destination="@id/fragment_my_report_history" />
     </fragment>
 
     <fragment
-        android:id="@+id/fragment_new_witness_report"
-        android:name="com.example.findu.presentation.ui.report.witness.NewWitnessReportFragment"
-        android:label="New Witness Report"
-        tools:layout="@layout/fragment_new_witness_report">
+        android:id="@+id/fragment_witness_report"
+        android:name="com.example.findu.presentation.ui.report.witness.WitnessReportFragment"
+        android:label="Witness Report"
+        tools:layout="@layout/fragment_witness_report">
         <action
-            android:id="@+id/action_fragment_new_witness_report_to_fragment_home"
+            android:id="@+id/action_fragment_witness_report_to_fragment_home"
             app:destination="@id/fragment_home"
             app:popUpToInclusive="true" />
         <action
-            android:id="@+id/action_fragment_new_witness_report_to_fragment_report_camera"
+            android:id="@+id/action_fragment_witness_report_to_fragment_report_camera"
             app:destination="@id/fragment_report_camera" />
         <action
-            android:id="@+id/action_fragment_new_witness_report_to_fragment_my_report_history"
+            android:id="@+id/action_fragment_witness_report_to_fragment_my_report_history"
             app:destination="@id/fragment_my_report_history" />
     </fragment>
